// web/src/shared/api/generated/model/studentAvatarConversationDetailResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AvatarEntryPoint } from './avatarEntryPoint';
import type { StudentAvatarMessageResult } from './studentAvatarMessageResult';

export interface StudentAvatarConversationDetailResult {
  id: string;
  entryPoint: AvatarEntryPoint;
  /** @nullable */
  subjectId: string | null;
  /** @nullable */
  subjectName: string | null;
  /** @nullable */
  lessonId: string | null;
  /** @nullable */
  lessonName: string | null;
  /** @nullable */
  sessionId: string | null;
  /** @nullable */
  questionId: string | null;
  startedAt: string;
  lastMessageAt: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  messageCount: number | string;
  messages: StudentAvatarMessageResult[];
}